Authentication 기타작업[BMMS]

+++ cafeback/AroundCafe/src/main/java/com/example/demo/security/jwt/dto/JwtDto.java
@@ -3,22 +3,22 @@ package com.example.demo.security.jwt.dto;
 import lombok.Builder;
 import lombok.Getter;
 
-import java.util.List;
+import java.util.Date;
 
 @Getter
 public class JwtDto {
 
     private final String type = "Bearer";
     private final String accessToken;
-    private final Integer accessTokenExp;
+    private final Long accessTokenExp;
     private final String refreshToken;
-    private final Integer refreshTokenExp;
+    private final Long refreshTokenExp;
     private final String email;
     private final String nickname;
     private final String role;
 
     @Builder
-    public JwtDto(String accessToken, Integer accessTokenExp, String refreshToken, Integer refreshTokenExp, String email, String nickname, String role) {
+    public JwtDto(String accessToken, Long accessTokenExp, String refreshToken, Long refreshTokenExp, String email, String nickname, String role) {
         this.accessToken = accessToken;
         this.accessTokenExp = accessTokenExp;
         this.refreshToken = refreshToken;
